refactor(client): extract component tests to sibling file
